tests refactoring, fix issue with (#348)

+++ goods-partner-backend/src/main/java/com/goodspartner/entity/Route.java
@@ -34,7 +34,7 @@ public class Route {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "routes_sequence")
     @SequenceGenerator(name = "routes_sequence", sequenceName = "routes_sequence")
-    private Integer id; // TODO migrate to long
+    private Long id;
 
     @Enumerated(value = EnumType.STRING)
     private RouteStatus status;
