+++ goods-partner-backend/src/main/java/com/goodspartner/web/controller/response/RouteActionResponse.java
@@ -22,7 +22,7 @@ public class RouteActionResponse {
     private UUID deliveryId;
     private DeliveryStatus deliveryStatus;
 
-    private int routeId;
+    private long routeId;
     private RouteStatus routeStatus;
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm")
